requirements.txt
openai>=1.80.0
openai-agents>=0.0.16
python-dotenv>=1.0.0
